Rename PublishedKeyword to Keyword
