+++ java/code/src/com/redhat/rhn/domain/errata/AbstractErrata.java
@@ -20,7 +20,7 @@ package com.redhat.rhn.domain.errata;
 import com.redhat.rhn.domain.BaseDomainHelper;
 import com.redhat.rhn.domain.channel.Channel;
 import com.redhat.rhn.domain.errata.impl.PublishedErrataFile;
-import com.redhat.rhn.domain.errata.impl.PublishedKeyword;
+import com.redhat.rhn.domain.errata.impl.Keyword;
 import com.redhat.rhn.domain.org.Org;
 import com.redhat.rhn.domain.rhnpackage.Package;
 import com.redhat.rhn.frontend.struts.Selectable;
@@ -522,7 +522,7 @@ public abstract class AbstractErrata extends BaseDomainHelper implements
          * Bah... this stinks since a keyword is just a string, but we have to
          * set the created/modified fields in the db.
          */
-        Keyword k = new PublishedKeyword();
+        Keyword k = new Keyword();
         k.setKeyword(keywordIn);
         addKeyword(k);
         k.setErrata(this);
