+++ java/code/src/com/redhat/rhn/domain/errata/Errata.java
@@ -18,6 +18,7 @@
 package com.redhat.rhn.domain.errata;
 
 import com.redhat.rhn.domain.channel.Channel;
+import com.redhat.rhn.domain.errata.impl.Keyword;
 import com.redhat.rhn.domain.org.Org;
 import com.redhat.rhn.domain.rhnpackage.Package;
